Update ALCE tests, small changes
