+++ skactiveml/pool/_alce.py
@@ -1,13 +1,20 @@
+import warnings
+
 import numpy as np
+from joblib import Parallel
+from joblib import delayed
 from sklearn import clone
+from sklearn.base import BaseEstimator
 from sklearn.base import RegressorMixin
+from sklearn.isotonic import IsotonicRegression
+from sklearn.metrics import euclidean_distances
 from sklearn.neighbors import NearestNeighbors
 from sklearn.utils import check_array, check_random_state
+from sklearn.utils import check_symmetric
 
 from skactiveml.base import SingleAnnotPoolBasedQueryStrategy
-from skactiveml.pool._mdsp import MDSP
-from skactiveml.utils import rand_argmax, MISSING_LABEL, check_scalar
 from skactiveml.utils import check_cost_matrix, check_missing_label
+from skactiveml.utils import rand_argmax, MISSING_LABEL, check_scalar
 
 
 class ALCE(SingleAnnotPoolBasedQueryStrategy):
@@ -19,7 +26,8 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
     Parameters
     ----------
     base_regressor : sklearn regressor
-    C: array-like, shape (n_classes, n_classes), optional (default=None)
+    cost_matrix: array-like, shape (n_classes, n_classes),
+                 optional (default=None)
         Cost matrix with C[i,j] defining the cost of predicting class j for a
         sample with the actual class i. Only supported for least confident
         variant.
@@ -46,7 +54,7 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
 
     def __init__(self,
                  base_regressor,
-                 C,
+                 cost_matrix,
                  embed_dim=None,
                  missing_label=MISSING_LABEL,
                  classes=None,
@@ -55,7 +63,7 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
                  random_state=None):
         super().__init__(random_state=random_state)
         self.base_regressor = base_regressor
-        self.C = C
+        self.cost_matrix = cost_matrix
         self.embed_dim = embed_dim
         self.missing_label = missing_label
         self.classes = classes
@@ -92,20 +100,32 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
         X = check_array(X, force_all_finite=False)
         y = check_array(y, force_all_finite=False, ensure_2d=False)
 
+        # Check base regressor
         if not isinstance(self.base_regressor, RegressorMixin):
             raise TypeError("'base_regressor' must be an sklearn regressor")
+
+        # Check cost matrix
+        check_cost_matrix(self.cost_matrix, len(self.classes))
+
+        # Check classes
         if self.classes is None:
-            self.classes = np.arange(len(self.C))
-        check_cost_matrix(self.C, len(self.classes))
-        if self.embed_dim is None:
-            self.embed_dim = len(self.classes)
-        check_scalar(self.embed_dim, 'embed_dim', int, min_val=1)
-        check_missing_label(self.missing_label)
-        self.random_state = check_random_state(self.random_state)
+            self.classes = np.arange(len(self.cost_matrix))
         if not set(y) <= set(self.classes):
             raise ValueError("y has labels that are not contained in "
                              "'classes'")
 
+        # Check embedding dimension
+        if self.embed_dim is None:
+            self.embed_dim = len(self.classes)
+        check_scalar(self.embed_dim, 'embed_dim', int, min_val=1)
+
+        # Check missing label
+        check_missing_label(self.missing_label)
+
+        # Check random state
+        self.random_state = check_random_state(self.random_state)
+
+        # Update mds parameters
         mds_params = {
             'metric': False,
             'n_components': self.embed_dim,
@@ -118,10 +138,16 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
             'random_state': self.random_state
         }
         if self.mds_params is not None:
+            if type(self.mds_params) is not dict:
+                raise TypeError("'mds_params' must be a dictionary or None")
             mds_params.update(self.mds_params)
         self.mds_params = mds_params
+
+        # Update nearest neighbor parameters
         if self.nn_params is None:
             self.nn_params = {}
+        if type(self.nn_params) is not dict:
+            raise TypeError("'nn_params' must be a dictionary or None")
 
         regressors = [
             clone(self.base_regressor) for _ in range(self.embed_dim)
@@ -129,8 +155,8 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
         n_classes = len(self.classes)
 
         dissimilarities = np.zeros((2 * n_classes, 2 * n_classes))
-        dissimilarities[:n_classes, n_classes:] = self.C
-        dissimilarities[n_classes:, :n_classes] = self.C.T
+        dissimilarities[:n_classes, n_classes:] = self.cost_matrix
+        dissimilarities[n_classes:, :n_classes] = self.cost_matrix.T
 
         W = np.zeros((2 * n_classes, 2 * n_classes))
         W[:n_classes, n_classes:] = 1
@@ -156,3 +182,372 @@ class ALCE(SingleAnnotPoolBasedQueryStrategy):
             return query_indices, np.array([utilities])
         else:
             return query_indices
+
+
+"""
+Multi-dimensional Scaling Partial (MDSP)
+This module is modified from
+https://github.com/scikit-learn/scikit-learn/blob/14031f6/sklearn/manifold/mds.py
+by Kuan-Hao Huang.
+"""
+
+# author: Nelle Varoquaux <nelle.varoquaux@gmail.com>
+# Licence: BSD
+
+
+def _smacof_single_p(similarities, n_uq, metric=True, n_components=2,
+                     init=None, max_iter=300, verbose=0, eps=1e-3,
+                     random_state=None):
+    """
+    Computes multidimensional scaling using SMACOF algorithm
+    Parameters
+    ----------
+    n_uq
+    similarities: symmetric ndarray, shape [n * n]
+        similarities between the points
+    metric: boolean, optional, default: True
+        compute metric or nonmetric SMACOF algorithm
+    n_components: int, optional, default: 2
+        number of dimension in which to immerse the similarities
+        overwritten if initial array is provided.
+    init: {None or ndarray}, optional
+        if None, randomly chooses the initial configuration
+        if ndarray, initialize the SMACOF algorithm with this array
+    max_iter: int, optional, default: 300
+        Maximum number of iterations of the SMACOF algorithm for a single run
+    verbose: int, optional, default: 0
+        level of verbosity
+    eps: float, optional, default: 1e-6
+        relative tolerance w.r.t stress to declare converge
+    random_state: integer or numpy.RandomState, optional
+        The generator used to initialize the centers. If an integer is
+        given, it fixes the seed. Defaults to the global numpy random
+        number generator.
+    Returns
+    -------
+    X: ndarray (n_samples, n_components), float
+               coordinates of the n_samples points in a n_components-space
+    stress_: float
+        The final value of the stress (sum of squared distance of the
+        disparities and the distances for all constrained points)
+    n_iter : int
+        Number of iterations run.
+    """
+    similarities = check_symmetric(similarities, raise_exception=True)
+
+    n_samples = similarities.shape[0]
+    random_state = check_random_state(random_state)
+
+    W = np.ones((n_samples, n_samples))
+    W[:n_uq, :n_uq] = 0.0
+    W[n_uq:, n_uq:] = 0.0
+
+    V = -W
+    V[np.arange(len(V)), np.arange(len(V))] = W.sum(axis=1)
+    e = np.ones((n_samples, 1))
+
+    Vp = np.linalg.inv(V + np.dot(e, e.T) / n_samples) - \
+         np.dot(e, e.T) / n_samples
+
+    sim_flat = similarities.ravel()
+    sim_flat_w = sim_flat[sim_flat != 0]
+    if init is None:
+        # Randomly choose initial configuration
+        X = random_state.rand(n_samples * n_components)
+        X = X.reshape((n_samples, n_components))
+    else:
+        # overrides the parameter p
+        n_components = init.shape[1]
+        if n_samples != init.shape[0]:
+            raise ValueError("init matrix should be of shape (%d, %d)" %
+                             (n_samples, n_components))
+        X = init
+
+    old_stress = None
+    ir = IsotonicRegression()
+    for it in range(max_iter):
+        # Compute distance and monotonic regression
+        dis = euclidean_distances(X)
+
+        if metric:
+            disparities = similarities
+        else:
+            dis_flat = dis.ravel()
+            # similarities with 0 are considered as missing values
+            dis_flat_w = dis_flat[sim_flat != 0]
+
+            # Compute the disparities using a monotonic regression
+            disparities_flat = ir.fit_transform(sim_flat_w, dis_flat_w)
+            disparities = dis_flat.copy()
+            disparities[sim_flat != 0] = disparities_flat
+            disparities = disparities.reshape((n_samples, n_samples))
+            disparities *= np.sqrt((n_samples * (n_samples - 1) / 2)
+                                   / (disparities ** 2).sum())
+            disparities[similarities == 0] = 0
+
+        # Compute stress
+        _stress = (W.ravel() * (
+                    (dis.ravel() - disparities.ravel()) ** 2)).sum()
+        _stress /= 2
+
+        # Update X using the Guttman transform
+        dis[dis == 0] = 1e-5
+        ratio = disparities / dis
+        _B = - W * ratio
+        _B[np.arange(len(_B)), np.arange(len(_B))] += (W * ratio).sum(axis=1)
+
+        X = np.dot(Vp, np.dot(_B, X))
+
+        dis = np.sqrt((X ** 2).sum(axis=1)).sum()
+
+        if verbose >= 2:
+            print('it: %d, stress %s' % (it, _stress))
+        if old_stress is not None:
+            if (old_stress - _stress / dis) < eps:
+                if verbose:
+                    print(f'breaking at iteration {it} with stress {_stress}')
+                break
+        old_stress = _stress / dis
+
+    return X, _stress, it + 1
+
+
+def smacof_p(similarities, n_uq, metric=True, n_components=2, init=None,
+             n_init=8, n_jobs=1, max_iter=300, verbose=0, eps=1e-3,
+             random_state=None, return_n_iter=False):
+    """
+    Computes multidimensional scaling using SMACOF (Scaling by Majorizing a
+    Complicated Function) algorithm
+    The SMACOF algorithm is a multidimensional scaling algorithm: it minimizes
+    a objective function, the *stress*, using a majorization technique. The
+    Stress Majorization, also known as the Guttman Transform, guarantees a
+    monotone convergence of Stress, and is more powerful than traditional
+    techniques such as gradient descent.
+    The SMACOF algorithm for metric MDS can summarized by the following steps:
+    1. Set an initial start configuration, randomly or not.
+    2. Compute the stress
+    3. Compute the Guttman Transform
+    4. Iterate 2 and 3 until convergence.
+    The nonmetric algorithm adds a monotonic regression steps before computing
+    the stress.
+    Parameters
+    ----------
+    similarities : symmetric ndarray, shape (n_samples, n_samples)
+        similarities between the points
+    metric : boolean, optional, default: True
+        compute metric or nonmetric SMACOF algorithm
+    n_components : int, optional, default: 2
+        number of dimension in which to immerse the similarities
+        overridden if initial array is provided.
+    init : {None or ndarray of shape (n_samples, n_components)}, optional
+        if None, randomly chooses the initial configuration
+        if ndarray, initialize the SMACOF algorithm with this array
+    n_init : int, optional, default: 8
+        Number of time the smacof_p algorithm will be run with different
+        initialisation. The final results will be the best output of the
+        n_init consecutive runs in terms of stress.
+    n_jobs : int, optional, default: 1
+        The number of jobs to use for the computation. This works by breaking
+        down the pairwise matrix into n_jobs even slices and computing them in
+        parallel.
+        If -1 all CPUs are used. If 1 is given, no parallel computing code is
+        used at all, which is useful for debugging. For n_jobs below -1,
+        (n_cpus + 1 + n_jobs) are used. Thus for n_jobs = -2, all CPUs but one
+        are used.
+    max_iter : int, optional, default: 300
+        Maximum number of iterations of the SMACOF algorithm for a single run
+    verbose : int, optional, default: 0
+        level of verbosity
+    eps : float, optional, default: 1e-6
+        relative tolerance w.r.t stress to declare converge
+    random_state : integer or numpy.RandomState, optional
+        The generator used to initialize the centers. If an integer is
+        given, it fixes the seed. Defaults to the global numpy random
+        number generator.
+    return_n_iter : bool
+        Whether or not to return the number of iterations.
+    Returns
+    -------
+    X : ndarray (n_samples,n_components)
+        Coordinates of the n_samples points in a n_components-space
+    stress : float
+        The final value of the stress (sum of squared distance of the
+        disparities and the distances for all constrained points)
+    n_iter : int
+        The number of iterations corresponding to the best stress.
+        Returned only if `return_n_iter` is set to True.
+    Notes
+    -----
+    "Modern Multidimensional Scaling - Theory and Applications" Borg, I.;
+    Groenen P. Springer Series in Statistics (1997)
+    "Nonmetric multidimensional scaling: a numerical method" Kruskal, J.
+    Psychometrika, 29 (1964)
+    "Multidimensional scaling by optimizing goodness of fit to a nonmetric
+    hypothesis" Kruskal, J. Psychometrika, 29, (1964)
+    """
+
+    similarities = check_array(similarities)
+    random_state = check_random_state(random_state)
+
+    if hasattr(init, '__array__'):
+        init = np.asarray(init).copy()
+        if not n_init == 1:
+            warnings.warn(
+                'Explicit initial positions passed: '
+                'performing only one init of the MDS instead of %d'
+                % n_init)
+            n_init = 1
+
+    best_pos, best_stress = None, None
+
+    if n_jobs == 1:
+        for it in range(n_init):
+            pos, stress, n_iter_ = _smacof_single_p(
+                similarities, n_uq, metric=metric,
+                n_components=n_components, init=init,
+                max_iter=max_iter, verbose=verbose,
+                eps=eps, random_state=random_state)
+            if best_stress is None or stress < best_stress:
+                best_stress = stress
+                best_pos = pos.copy()
+                best_iter = n_iter_
+    else:
+        seeds = random_state.randint(np.iinfo(np.int32).max, size=n_init)
+        results = Parallel(n_jobs=n_jobs, verbose=max(verbose - 1, 0))(
+            delayed(_smacof_single_p)(
+                similarities, n_uq, metric=metric, n_components=n_components,
+                init=init, max_iter=max_iter, verbose=verbose, eps=eps,
+                random_state=seed)
+            for seed in seeds)
+        positions, stress, n_iters = zip(*results)
+        best = np.argmin(stress)
+        best_stress = stress[best]
+        best_pos = positions[best]
+        best_iter = n_iters[best]
+
+    if return_n_iter:
+        return best_pos, best_stress, best_iter
+    else:
+        return best_pos, best_stress
+
+
+class MDSP(BaseEstimator):
+    """Multidimensional scaling
+    Parameters
+    ----------
+    metric : boolean, optional, default: True
+        compute metric or nonmetric SMACOF (Scaling by Majorizing a
+        Complicated Function) algorithm
+    n_components : int, optional, default: 2
+        number of dimension in which to immerse the similarities
+        overridden if initial array is provided.
+    n_init : int, optional, default: 4
+        Number of time the smacof_p algorithm will be run with different
+        initialisation. The final results will be the best output of the
+        n_init consecutive runs in terms of stress.
+    max_iter : int, optional, default: 300
+        Maximum number of iterations of the SMACOF algorithm for a single run
+    verbose : int, optional, default: 0
+        level of verbosity
+    eps : float, optional, default: 1e-6
+        relative tolerance w.r.t stress to declare converge
+    n_jobs : int, optional, default: 1
+        The number of jobs to use for the computation. This works by breaking
+        down the pairwise matrix into n_jobs even slices and computing them in
+        parallel.
+        If -1 all CPUs are used. If 1 is given, no parallel computing code is
+        used at all, which is useful for debugging. For n_jobs below -1,
+        (n_cpus + 1 + n_jobs) are used. Thus for n_jobs = -2, all CPUs but one
+        are used.
+    random_state : integer or numpy.RandomState, optional
+        The generator used to initialize the centers. If an integer is
+        given, it fixes the seed. Defaults to the global numpy random
+        number generator.
+    dissimilarity : string
+        Which dissimilarity measure to use.
+        Supported are 'euclidean' and 'precomputed'.
+    Attributes
+    ----------
+    embedding_ : array-like, shape [n_components, n_samples]
+        Stores the position of the dataset in the embedding space
+    stress_ : float
+        The final value of the stress (sum of squared distance of the
+        disparities and the distances for all constrained points)
+    References
+    ----------
+    "Modern Multidimensional Scaling - Theory and Applications" Borg, I.;
+    Groenen P. Springer Series in Statistics (1997)
+    "Nonmetric multidimensional scaling: a numerical method" Kruskal, J.
+    Psychometrika, 29 (1964)
+    "Multidimensional scaling by optimizing goodness of fit to a nonmetric
+    hypothesis" Kruskal, J. Psychometrika, 29, (1964)
+    """
+
+    def __init__(self, n_components=2, n_uq=1, metric=True, n_init=4,
+                 max_iter=300, verbose=0, eps=1e-3, n_jobs=1,
+                 random_state=None, dissimilarity="euclidean"):
+        self.n_components = n_components
+        self.n_uq = n_uq
+        self.dissimilarity = dissimilarity
+        self.metric = metric
+        self.n_init = n_init
+        self.max_iter = max_iter
+        self.eps = eps
+        self.verbose = verbose
+        self.n_jobs = n_jobs
+        self.random_state = random_state
+
+    @property
+    def _pairwise(self):
+        return self.kernel == "precomputed"
+
+    def fit(self, X, y=None, init=None):
+        """
+        Computes the position of the points in the embedding space
+        Parameters
+        ----------
+        X : array, shape=[n_samples, n_features], or [n_samples, n_samples] \
+                if dissimilarity='precomputed'
+            Input data.
+        init : {None or ndarray, shape (n_samples,)}, optional
+            If None, randomly chooses the initial configuration
+            if ndarray, initialize the SMACOF algorithm with this array.
+        """
+        self.fit_transform(X, init=init)
+        return self
+
+    def fit_transform(self, X, y=None, init=None):
+        """
+        Fit the data from X, and returns the embedded coordinates
+        Parameters
+        ----------
+        X : array, shape=[n_samples, n_features], or [n_samples, n_samples] \
+                if dissimilarity='precomputed'
+            Input data.
+        init : {None or ndarray, shape (n_samples,)}, optional
+            If None, randomly chooses the initial configuration
+            if ndarray, initialize the SMACOF algorithm with this array.
+        """
+        X = check_array(X)
+        if X.shape[0] == X.shape[1] and self.dissimilarity != "precomputed":
+            warnings.warn("The MDS API has changed. ``fit`` now constructs an"
+                          " dissimilarity matrix from data. To use a custom "
+                          "dissimilarity matrix, set "
+                          "``dissimilarity=precomputed``.")
+
+        if self.dissimilarity == "precomputed":
+            self.dissimilarity_matrix_ = X
+        elif self.dissimilarity == "euclidean":
+            self.dissimilarity_matrix_ = euclidean_distances(X)
+        else:
+            raise ValueError("Proximity must be 'precomputed' or 'euclidean'."
+                             " Got %s instead" % str(self.dissimilarity))
+
+        self.embedding_, self.stress_, self.n_iter_ = smacof_p(
+            self.dissimilarity_matrix_, self.n_uq, metric=self.metric,
+            n_components=self.n_components, init=init, n_init=self.n_init,
+            n_jobs=self.n_jobs, max_iter=self.max_iter, verbose=self.verbose,
+            eps=self.eps, random_state=self.random_state,
+            return_n_iter=True)
+
+        return self.embedding_
